Move Link component to container level so data layer card doesn't have a dependency on it
